refactor(PR-746): move pose-component from joint-control to main menu
