+++ src/app/app.component.html
@@ -43,6 +43,29 @@
                                 class="nav-item"
                                 role="presentation">
                                 <a
+                                    id="pose-nav"
+                                    class="nav-link sidebar-btn"
+                                    routerLinkActive="btn-active"
+                                    routerLink="pose">
+                                    <svg
+                                        class="mr-1"
+                                        height="18px"
+                                        xmlns="http://www.w3.org/2000/svg"
+                                        viewBox="0 0 55.186 55.186">
+                                        <circle
+                                            cx="27.906"
+                                            cy="4.645"
+                                            r="4.645" />
+                                        <path
+                                            d="M46.486,11.079c-5.092,0-10.188,0-15.279,0c-1.959-1.276-4.733-1.28-6.672,0 c-5.278,0-10.558,0-15.836,0c-2.994,0-2.994,4.642,0,4.642c4.522,0,9.044,0,13.566,0c0,4.435,0,8.87,0,13.306 c0,0.112,0.011,0.219,0.016,0.33c-0.002,0.053-0.015,0.098-0.015,0.152c0,7.706-0.014,15.412-0.332,23.113 c-0.141,3.426,5.168,3.412,5.309,0c0.25-6.036,0.311-12.075,0.326-18.115c0.252,0.016,0.504,0.018,0.755,0.002 c0.016,6.039,0.077,12.078,0.326,18.113c0.141,3.412,5.449,3.426,5.309,0c-0.318-7.701-0.332-15.407-0.332-23.113 c0-0.236-0.029-0.455-0.078-0.659c0-4.376,0-8.752,0-13.129c4.313,0,8.629,0,12.939,0C49.48,15.721,49.48,11.079,46.486,11.079z" />
+                                    </svg>
+                                    Poses
+                                </a>
+                            </li>
+                            <li
+                                class="nav-item"
+                                role="presentation">
+                                <a
                                     id="camera-nav"
                                     class="nav-link sidebar-btn"
                                     routerLinkActive="btn-active"
