+++ src/app/app.module.ts
@@ -38,7 +38,7 @@ import {PythonCodeComponent} from "./program/program-splitscreen/python-code/pyt
 import {ConsoleComponent} from "./program/program-splitscreen/console/console.component";
 import {SmartConnectComponent} from "./ui-components/smart-connect/smart-connect.component";
 import {NgOptimizedImage} from "@angular/common";
-import {PoseComponent} from "./joint-control/pose/pose.component";
+import {PoseComponent} from "./pose/pose.component";
 import {MarkdownModule} from "ngx-markdown";
 
 @NgModule({
