+++ src/app/joint-control/joint-control.component.html
@@ -11,16 +11,6 @@
                 {{joint.label}}
             </a>
         </li>
-        <li
-            class="nav-item"
-            routerLinkActive="navbar-btn-active">
-            <a
-                id="poses-nav"
-                class="nav-link sidebar-btn"
-                routerLink="pose">
-                Poses
-            </a>
-        </li>
     </ul>
     <div class="mt-3">
         <div class="background-image">
