--- src/app/joint-control/pose/pose.component.html
@@ -1,125 +1,0 @@
-<div class="d-flex flex-row m-0 pb-2 bt-0 pl-1 pr-1">
-    <div class="flex-grow-1 pl-2">Poses</div>
-    <button
-        type="button"
-        id="save-pose-button"
-        class="btn btn-regular"
-        (click)="savePose()">
-        Save current pose
-    </button>
-</div>
-<div
-    *ngFor="let pose of poses | async"
-    class="pose-item d-flex flex-row mt-1 mb-2 pl-1 pr-1"
-    [ngClass]="pose.poseId == selectedPoseId ? 'selected' : 'unselected'">
-    <button
-        class="apply-button flex-shrink-0"
-        [disabled]="!pose.active"
-        (click)="applyPose(pose)">
-        <img
-            class="apply-pose"
-            src="../../../assets/joint-control/other/apply-pose.svg"
-            alt="apply-pose-image" />
-    </button>
-    <button
-        class="pose-label flex-grow-1 flex-shrink-1 pl-3"
-        (click)="selectPose(pose)"
-        (dblclick)="renamePose(pose)">
-        {{pose.name}}
-    </button>
-    <button
-        #renameButton
-        class="rename-button flex-shrink-0"
-        (click)="renamePose(pose)">
-        <img
-            class="rename-icon"
-            src="../../../assets/edit.svg"
-            alt="rename-pose-image" />
-    </button>
-    <button
-        class="delete-button flex-shrink-0"
-        (click)="deletePose(pose)">
-        <img
-            class="delete-icon"
-            src="../../../assets/delete.svg"
-            alt="delete-pose-image" />
-    </button>
-</div>
-
-<ng-template
-    #modalContent
-    let-modal
-    id="template">
-    <div
-        class="row display-flex modal-header"
-        id="modal-header">
-        <div
-            class="col-10"
-            id="header-title">
-            <span id="header-title-text">{{modalTitle}}</span>
-        </div>
-        <div
-            class="col-2 text-right"
-            id="close-button">
-            <button
-                id="modal-close-button"
-                type="button"
-                (click)="modal.dismiss('Cross')"
-                class="btn">
-                <svg
-                    xmlns="http://www.w3.org/2000/svg"
-                    width="10"
-                    height="16">
-                    <text
-                        font-family="Titillium"
-                        fill="#FFF"
-                        font-weight="bold"
-                        font-size="16"
-                        y="11.062">
-                        X
-                    </text>
-                </svg>
-            </button>
-        </div>
-    </div>
-    <div
-        class="row m-2 mt-4 pr-2"
-        id="row-name">
-        <label
-            for="name-input"
-            class="col-3 mt-2 justify-content-end"
-            id="label-name"
-            >Name:</label
-        >
-        <input
-            ngbAutofocus
-            class="col-9"
-            type="text"
-            (keydown)="($event).key ==='Enter' ? modal.close('Save'): ''"
-            name="name-input"
-            [formControl]="nameFormControl"
-            id="name-input" />
-    </div>
-    <div class="row justify-content-center mt-2 mb-3">
-        <div class="col-3 align-self-center">
-            <button
-                type="button"
-                class="btn btn-regular"
-                id="modal-save-button"
-                [disabled]="nameFormControl.invalid"
-                (keydown.enter)="modal.close('Save')"
-                (click)="modal.close('Save')">
-                Save
-            </button>
-        </div>
-        <div class="col-3 align-self-center">
-            <button
-                type="button"
-                class="btn btn-regular"
-                id="modal-cancel-button"
-                (click)="modal.dismiss('Cancel')">
-                Cancel
-            </button>
-        </div>
-    </div>
-</ng-template>
