+++ src/app/pose/pose.component.css
