+++ src/app/pose/pose.component.spec.ts
